+++ packages/stream_chat/lib/src/core/models/message.g.dart
@@ -92,6 +92,9 @@ Message _$MessageFromJson(Map<String, dynamic> json) => Message(
       draft: json['draft'] == null
           ? null
           : Draft.fromJson(json['draft'] as Map<String, dynamic>),
+      reminder: json['reminder'] == null
+          ? null
+          : MessageReminder.fromJson(json['reminder'] as Map<String, dynamic>),
     );
 
 Map<String, dynamic> _$MessageToJson(Message instance) => <String, dynamic>{
@@ -110,5 +113,6 @@ Map<String, dynamic> _$MessageToJson(Message instance) => <String, dynamic>{
       if (instance.restrictedVisibility case final value?)
         'restricted_visibility': value,
       'draft': instance.draft?.toJson(),
+      'reminder': instance.reminder?.toJson(),
       'extra_data': instance.extraData,
     };